move vpromms table row and column header to separate components
